refactor(kfx): extract host-agnostic planKfx from the gui loader
Split loadKfx's discovery + trust/tier decision into planKfx in
@kungfu-tech/kfx, with fs/path/crypto injected, returning a neutral load
plan (KfxPlanEntry) that instantiates no View/DOM/Electron. The gui loader
now consumes planKfx and only lands the View in this renderer; KfxEntry
becomes KfxPlanEntry & { View }.

Both hosts can now import one load rule and reach an identical verdict for
the same kfx, instead of the rule being welded to the renderer.
